requirements.txt
Flask
numpy
Pillow
pycryptodome
gunicorn
